prepare version 2.4.0

+++ src/px_types.h
@@ -56,7 +56,7 @@ typedef struct px_config_t {
     apr_thread_cond_t *health_check_cond;
     int px_errors_threshold;
     volatile apr_uint32_t px_errors_count;
-    int health_check_interval; // in ms
+    long health_check_interval; // in ms
 } px_config;
 
 typedef struct health_check_data_t {
